chore: move generateCreate operation to utils
