+++ packages/did-method-photon/src/test/utils.ts
@@ -13,31 +13,32 @@
  */
 
 import { MongoDb } from '@sidetree/db';
-import { ICas } from '@sidetree/common';
+import { methods } from '@sidetree/wallet';
+import { ICas, PublicKeyPurpose } from '@sidetree/common';
 import QLDBLedger from '@sidetree/qldb';
 import { S3Cas } from '@sidetree/cas-s3';
 import Photon from '../Photon';
 import config from './photon-config.json';
 
-const resetDatabase = async (): Promise<void> => {
+export const resetDatabase = async (): Promise<void> => {
   await MongoDb.resetDatabase(
     config.mongoDbConnectionString,
     config.databaseName!
   );
 };
 
-const getTestLedger = async (): Promise<QLDBLedger> => {
+export const getTestLedger = async (): Promise<QLDBLedger> => {
   const ledger = new QLDBLedger(config.qldbLedger, config.qldbLedgerTable);
   await ledger.reset();
   return ledger;
 };
 
-const getTestCas = (): ICas => {
+export const getTestCas = (): ICas => {
   const cas = new S3Cas(config.s3BucketName);
   return cas;
 };
 
-const getTestPhoton = async (): Promise<Photon> => {
+export const getTestPhoton = async (): Promise<Photon> => {
   await resetDatabase();
   const ledger = await getTestLedger();
   const cas = await getTestCas();
@@ -46,7 +47,7 @@ const getTestPhoton = async (): Promise<Photon> => {
   return photon;
 };
 
-const replaceMethod = (
+export const replaceMethod = (
   result: JSON,
   defaultMethod = 'did:elem',
   specificMethod = 'did:photon'
@@ -60,10 +61,35 @@ const replaceMethod = (
   return updateResult;
 };
 
-export {
-  resetDatabase,
-  getTestLedger,
-  getTestCas,
-  getTestPhoton,
-  replaceMethod,
+export const generateCreateOperation = async (publicKey: any): Promise<any> => {
+  // We could generate the create operation like this
+  /*
+  const mnemonic = crypto.mnemonic.mnemonic[0];
+  const createOperation = await methods.getCreateOperationForProfile(
+    mnemonic,
+    i
+  );
+  */
+  // However this is too slow because it generates new keys for every create
+  // operation which cause the tests to timeout for batch size larger than 1000
+
+  // Therefore for the purpose of showing the we can process large batches
+  // we will generate create operation for did documents that share the same key
+  const documentModel = {
+    public_keys: [
+      {
+        // id is random so that each id (and therefore each did) is different
+        id: Math.random(),
+        type: 'JsonWebKey2020',
+        jwk: publicKey,
+        purpose: [PublicKeyPurpose.General],
+      },
+    ],
+  };
+  const createOperation = await methods.getCreatePayloadFromDocumentModel(
+    documentModel,
+    publicKey,
+    publicKey
+  );
+  return createOperation;
 };
